requirements.txt
discord.py==2.4.0
disckit==0.8
python-dotenv==1.0.1